feat: Update DMS provider to expect multiple sortable properties.
The `sortby` mechanism in the query method has been updated to expect sortable properties in a collection rather than as a single item, enabling multiple sortable properties. This has been done by adding a new `sortables` list member to the `MSCDMSCoreAPIProvider` class. This attribute is set to a `sortables` definition if it exists as a list in `msc-pygeoapi-config.yml`. Otherwise, it stores just the `time_field`.